Internalized Integer/Float
This makes it impossible to instantiate an Integer in such a way that
it's not represented by its smallest version. This was an assumption
that all of the Integer::as_ methods contained, and it was either going
to be an exercise of updating those methods or I could handle compacting
on construction.

Floats don't currently try to demote from f64 to f32 automatically. The
float conversion methods didn't have the same assumption, making the
extra work truly extra work for no real gains.