Changed Read-based functions to DeserializedOwned
This was an oversight in my initial implementation. There's no way to
actually borrow bytes from an arbitrary reader -- the bytes would be
borrowed from the 'de context, not the type's lifetime.

Fixing this unfortunately lowers the code coverage, as I can't trigger
one of the error cases anymore. I'm leaving the code there instead of
changing it to unreachable in case it is still reachable in ways I
couldn't figure out.