Implemented preshared symbol maps
When I last left off, I wanted to finish my vision for symbol maps.
Unfortunately, it required a bit more work than I had anticipated.
Specifically, the serializing side needed to be refactored to support
storing the symbols and the prepopulate function required custom
serialization.

With these changes, Pot can be initialized using a preshared list of
symbols, or a persistently updated SymbolMap can be used for
long-running and resumable transmissions.

The only remaining feature might be a way to configure serialization so
that it emits all symbols after a specific ID, allowing one party to
send to another party who has a partial dictionary.