Removing serde's derive feature
This set of changes started as an experiment in removing the derive
feature from serde after having seen some interesting analysis of build
parallelization due to crates usage of the derive macro.

In the process of this, the examples started looking ugly due to the
split `use` statements for the two crates. I realized one area of API
cleanup I hadn't done yet was make more ergonomic serialize/deserialize
functions for the SymbolMaps. After I did this, the examples no longer
needed the traits, and only needed the derive macros.