Added Compatibility + deserialize_any fix
Closes #11

This set of changes adds new compatibility options that ensure no
breaking changes in v3.0, but allow users to opt into the new format
change that will become default in v4.0.

The issue that this new format change fixes is that an enum variant with
associated data and an enum variant with no assicated data were both
encoded using a `named` atom. In `deserialize_any` we have no way to
know whether the next atom after the name is actually data associated
with the enum variant or if it's the next value in the stream.

The new format encodes enum variants with no associated data as a symbol
atom only. Parsers before v3.0.1 will not be able to read this change.